requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.11
cycler==0.11.0
fonttools==4.29.1
idna==3.3
imageio==2.14.1
kiwisolver==1.3.2
networkx==2.6.3
numpy==1.22.1
packaging==21.3
Pillow==9.0.1
pkg_resources==0.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
requests==2.27.1
six==1.16.0
tifffile==2021.11.2
torch==1.10.2
torchtext==0.11.2
torchvision==0.11.3
tornado==6.1
tqdm==4.62.3
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
